requirements.txt
certifi==2022.12.7
setuptools==65.5.0
wheel==0.37.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
einops==0.6.0
entrypoints==0.4
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
gitdb==4.0.10
gitpython==3.1.31
gradio==3.22.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.2
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-loader==0.1
lightning-utilities==0.8.0
linkify-it-py==2.0.0
lit==15.0.7
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
open-clip-torch==2.16.0
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.8.7
packaging==23.0
pandas==1.5.3
pillow==9.3.0
prettytable==3.6.0
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.6
pydeck==0.8.0
pydub==0.25.1
pygments==2.14.0
pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywavelets==1.4.1
pyyaml==6.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
rich==13.3.2
safetensors==0.3.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
streamlit==1.20.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.8.15.dev0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torchmetrics==0.11.4
tornado==6.2
tqdm==4.65.0
transformers==4.27.1
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.21.1
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
websockets==10.4
yarl==1.8.2
zipp==3.15.0